+++ tests/RequestHandler/OrderRequestHandlerTest.php
@@ -25,6 +25,7 @@ use DateTime;
 use Exception;
 use MintWare\Streams\MemoryStream;
 use PHPUnit\Framework\TestCase;
+use Psr\Http\Message\RequestInterface;
 
 class OrderRequestHandlerTest extends TestCase
 {
@@ -62,7 +63,7 @@ class OrderRequestHandlerTest extends TestCase
         $this->assertEquals(1, $data['paging']['totalPages']);
         $this->assertEquals(1, $data['paging']['totalCount']);
         $body = '{"paging":{"page":1,"totalCount":1,"totalPages":1},"orders":[{"order_id":"1234","order_number":"456","currency_code":"EUR","delivery_source_country_code":"DE","nick_name":"GirlWhoCanFly","ship_cost":4.9,"invoice_address":{"firstname":"Kara","lastname":"Zor-El","street":"Argo Street","housenumber":"1022","address2":"Window","postcode":"90012","city":"National City","country_code":"US","company":"D.E.O.","state":"CA"},"delivery_address":{"firstname":"Kara","lastname":"Zor-El","street":"Argo Street","housenumber":"1022","address2":"Window","postcode":"90012","city":"National City","country_code":"US","company":"D.E.O.","state":"CA"},"order_date":"2019-01-01T20:00:15+0000","email":"secret@deo.tld","phone1":"0123456789","pay_date":"2019-01-01T23:00:15+0000","ship_date":"2019-01-02T02:00:15+0000","payment_method":1,"order_status_id":2,"seller_comment":"Psst","shippingprofile_id":"super-fast","vat_id":"DE-123456","payment_transaction_id":"123444"}]}';
-        $this->assertEquals($body, $response->getBody());
+        $this->assertEquals($body, (string)$response->getBody());
     }
 
     public function testGetOrdersFailsInvalidDAte()
@@ -103,10 +104,11 @@ class OrderRequestHandlerTest extends TestCase
              ->willThrowException(new OrderNotFoundException());
         $handler = new OrderRequestHandler($repo);
 
+        /** @var RequestInterface $request */
         $request = new Request();
         $uri = new Uri('http://localhost/?Action=AckOrder');
         $req = $request->withUri($uri)
-                       ->withBody(new MemoryStream(json_encode(['OrderId' => '1'])));
+                       ->withBody(new MemoryStream(http_build_query(['OrderId' => '1'])));
 
         $response = $handler->handle($req, ['Action' => 'AckOrder']);
 
@@ -121,10 +123,11 @@ class OrderRequestHandlerTest extends TestCase
              ->willThrowException(new Exception('Unknown Error'));
         $handler = new OrderRequestHandler($repo);
 
+        /** @var RequestInterface $request */
         $request = new Request();
         $uri = new Uri('http://localhost/?Action=AckOrder&OrderId=1');
         $req = $request->withUri($uri)
-                       ->withBody(new MemoryStream(json_encode(['OrderId' => '1'])));
+                       ->withBody(new MemoryStream(http_build_query(['OrderId' => '1'])));
 
 
         $response = $handler->handle($req, ['Action' => 'AckOrder']);
@@ -140,10 +143,11 @@ class OrderRequestHandlerTest extends TestCase
              ->willReturn(null);
         $handler = new OrderRequestHandler($repo);
 
+        /** @var RequestInterface $request */
         $request = new Request();
         $uri = new Uri('http://localhost/?Action=AckOrder&OrderId=1');
         $req = $request->withUri($uri)
-                       ->withBody(new MemoryStream(json_encode(['OrderId' => '1'])));
+                       ->withBody(new MemoryStream(http_build_query(['OrderId' => '1'])));
 
 
         $response = $handler->handle($req, ['Action' => 'AckOrder']);
@@ -234,19 +238,20 @@ class OrderRequestHandlerTest extends TestCase
         $this->assertEquals('Es wurde keine OrderId übergeben', (string)$response->getBody());
     }
 
-    public function testSetOrderStateFailsNoNewStateId()
+    public function testSetOrderStateFailsNoNewStateTypeId()
     {
         $repo = $this->createMock(OrdersRepositoryInterface::class);
         $handler = new OrderRequestHandler($repo);
 
+        /** @var RequestInterface $request */
         $request = new Request();
         $uri = new Uri('http://localhost/?Action=SetOrderState');
-        $req = $request->withUri($uri)->withBody(new MemoryStream(json_encode(['OrderId' => '1'])));
+        $req = $request->withUri($uri)->withBody(new MemoryStream(http_build_query(['OrderId' => '1'])));
 
         $response = $handler->handle($req, ['Action' => 'SetOrderState']);
 
         $this->assertEquals(400, $response->getStatusCode());
-        $this->assertEquals('Es wurde keine NewStateId übergeben', (string)$response->getBody());
+        $this->assertEquals('Es wurde keine NewStateTypeId übergeben', (string)$response->getBody());
     }
 
     public function testSetOrderStateFailsNotFound()
@@ -256,9 +261,10 @@ class OrderRequestHandlerTest extends TestCase
              ->willThrowException(new OrderNotFoundException());
         $handler = new OrderRequestHandler($repo);
 
+        /** @var RequestInterface $request */
         $request = new Request();
         $uri = new Uri('http://localhost/?Action=SetOrderState');
-        $req = $request->withUri($uri)->withBody(new MemoryStream(json_encode(['OrderId' => 1, 'NewStateId' => 1])));
+        $req = $request->withUri($uri)->withBody(new MemoryStream(http_build_query(['OrderId' => 1, 'NewStateTypeId' => 1])));
 
         $response = $handler->handle($req, ['Action' => 'SetOrderState']);
 
@@ -273,9 +279,10 @@ class OrderRequestHandlerTest extends TestCase
              ->willThrowException(new Exception('Unknown Error'));
         $handler = new OrderRequestHandler($repo);
 
+        /** @var RequestInterface $request */
         $request = new Request();
         $uri = new Uri('http://localhost/?Action=SetOrderState');
-        $req = $request->withUri($uri)->withBody(new MemoryStream(json_encode(['OrderId' => 1, 'NewStateId' => 1])));
+        $req = $request->withUri($uri)->withBody(new MemoryStream(http_build_query(['OrderId' => 1, 'NewStateTypeId' => 1])));
 
         $response = $handler->handle($req, ['Action' => 'SetOrderState', 'OrderId' => '1']);
 
@@ -290,9 +297,10 @@ class OrderRequestHandlerTest extends TestCase
              ->willReturn(true);
         $handler = new OrderRequestHandler($repo);
 
+        /** @var RequestInterface $request */
         $request = new Request();
         $uri = new Uri('http://localhost/?Action=SetOrderState');
-        $req = $request->withUri($uri)->withBody(new MemoryStream(json_encode(['OrderId' => 1, 'NewStateId' => 1])));
+        $req = $request->withUri($uri)->withBody(new MemoryStream(http_build_query(['OrderId' => 1, 'NewStateTypeId' => 1])));
 
         $response = $handler->handle($req, ['Action' => 'SetOrderState', 'OrderId' => '1']);